requirements_fixed.txt
# Requirements para L7Nutri - Versão Simplificada Render
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
